+++ src/include/DataCollector.hpp
@@ -235,7 +235,7 @@ namespace DCollector
          */
         virtual void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
+                size_t count,
                 const char *name,
                 const void *data) = 0;
 
@@ -258,9 +258,9 @@ namespace DCollector
          */
         virtual void append(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
-                uint32_t offset,
-                uint32_t stride,
+                size_t count,
+                size_t offset,
+                size_t stride,
                 const char *name,
                 const void *data) = 0;
