+++ src/include/ParallelDomainCollector.hpp
@@ -174,7 +174,7 @@ namespace DCollector
 
         void appendDomain(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
+                size_t count,
                 const char *name,
                 const Dimensions domainOffset,
                 const Dimensions domainSize,
@@ -182,9 +182,9 @@ namespace DCollector
 
         void appendDomain(int32_t id,
                 const CollectionType& type,
-                uint32_t count,
-                uint32_t offset,
-                uint32_t striding,
+                size_t count,
+                size_t offset,
+                size_t striding,
                 const char *name,
                 const Dimensions domainOffset,
                 const Dimensions domainSize,
